+++ testdata/import-alias-naming-custom-config.go
